+++ projects/frontend/es-app/src/guis/menus/GuiMenuGamelistOptions.cpp
@@ -1,5 +1,7 @@
 #include "GuiMenuGamelistOptions.h"
 #include "guis/GuiDownloader.h"
+#include "guis/SearchForceOptions.h"
+#include "GuiCheckMenu.h"
 #include <guis/GuiSearch.h>
 #include <MainRunner.h>
 #include <views/gamelist/ISimpleGameListView.h>
@@ -28,6 +30,7 @@ GuiMenuGamelistOptions::GuiMenuGamelistOptions(WindowManager& window, SystemData
 
     if(!mGamelist.getCursor()->IsEmpty())
     {
+      mAlias = mGamelist.getCursor()->Metadata().Alias();
       if (!mSystem.IsVirtual() && mGamelist.getCursor()->IsGame() && !mGamelist.getCursor()->TopAncestor().ReadOnly() &&
           !mSystem.IsScreenshots())
       {
@@ -40,8 +43,17 @@ GuiMenuGamelistOptions::GuiMenuGamelistOptions(WindowManager& window, SystemData
       {
         AddSubMenu(_("DELETE SCREENSHOT"), (int) Components::DeleteScreeshot);
       }
+
+      if(!mGamelist.getCursor()->Metadata().Alias().empty())
+        AddSubMenu(_("SEARCH OTHERS VERSIONS"), (int) Components::SearchSiblings, _(MENUMESSAGE_GAMELISTOPTION_SEARCH_SIBLINGS_MSG));
     }
+    if(!mGamelist.getCursor()->Metadata().Families().empty())
+      AddSubMenu(_("SEARCH GAMES OF SAME LICENCE"), (int) Components::SearchFamily, _(MENUMESSAGE_GAMELISTOPTION_SEARCH_LICENCE_MSG));
   }
+
+  // open search wheel for this system
+  if (!system.IsFavorite())
+    AddSubMenu(_("SEARCH GAMES HERE"),  (int)Components::Search, Strings::Empty);
 
   RefreshGameMenuContext();
 
@@ -230,6 +242,7 @@ void GuiMenuGamelistOptions::OptionListComponentChanged(int id, int index, const
 
 void GuiMenuGamelistOptions::SubMenuSelected(int id)
 {
+  Strings::Vector families = mGamelist.getCursor()->Metadata().FamiliesAsList();
   switch((Components)id)
   {
     case Components::Download:
@@ -278,6 +291,42 @@ void GuiMenuGamelistOptions::SubMenuSelected(int id)
       mWindow.pushGui(new GuiSearch(mWindow, mSystemManager));
       break;
     }
+    case Components::SearchSiblings:
+    {
+      std::string alias = mGamelist.getCursor()->Metadata().Alias();
+      SearchForcedOptions forcedOptions = SearchForcedOptions(alias, FolderData::FastSearchContext::Alias, true);
+      mWindow.pushGui(new GuiSearch(mWindow, mSystemManager, &forcedOptions));
+      break;
+    }
+
+    case Components::SearchFamily:
+      if (families.size() == 1)
+      {
+        std::string family = mGamelist.getCursor()->Metadata().Families();
+        SearchForcedOptions forcedOptions = SearchForcedOptions(family, FolderData::FastSearchContext::Family, true);
+        mWindow.pushGui(new GuiSearch(mWindow, mSystemManager, &forcedOptions));
+      } else
+      {
+        std::list<ButtonComponent> buttons;
+        for (auto& family : families)
+        {
+          buttons.push_back(ButtonComponent(mWindow,
+                            family,
+                            family,
+                            [this, family] {
+            std::string familyLocal = family;
+            SearchForcedOptions forcedOptions = SearchForcedOptions(familyLocal, FolderData::FastSearchContext::Family, true);
+            mWindow.pushGui(new GuiSearch(mWindow, mSystemManager, &forcedOptions));}));
+        }
+        mWindow.pushGui(new GuiCheckMenu(mWindow,
+                                         _("licences"),
+                                         mGamelist.getCursor()->Name(),
+                                         0,
+                                         buttons
+                                        ));
+      }
+      break;
+
     case Components::JumpToLetter:
     case Components::Sorts:
     case Components::Regions:
@@ -302,6 +351,8 @@ void GuiMenuGamelistOptions::SwitchComponentChanged(int id, bool status)
     case Components::UpdateGamelist:
     case Components::Delete:
     case Components::DeleteScreeshot:
+    case Components::SearchSiblings:
+    case Components::SearchFamily:
     case Components::Quit: break;
   }